+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Support static labels per pipeline to tag uploaded records, [PR-24](https://github.com/reductstore/reductstore_agent/pull/24)
+
 ## [0.1.0] - 2025-06-09
 
 ### Added